+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project tries to adhere to [Semantic Versioning](https://semver.org/spe
 ### Fixed
 - show choropleth on revisit of status quo
 - fix empty popup on doubleclick
+- keep detail panel open on slider value change
 
 ## [1.0.1] - 2023-11-01
 ### Added